Refactor tok builder into resource subclasses
